requirements.txt
Flask
mysql-connector-python
Flask-Mail
